+++ .users/amin/GameSetting.txt
@@ -1,4 +1,4 @@
 1
-0
-0
+1
+4
 3
