+++ temp.txt
@@ -1,0 +1,15 @@
+gcc Rogue.c -lncursesw -lSDL2 -lSDL2_mixer
+
+📊 پیشنهاد XP برای فعالیت‌های مختلف در Rogue
+فعالیت مقدار 					XP پیشنهادی				توضیح
+کشتن دشمن ضعیف (Goblin, Bat, Slime, etc.)		5 XP		دشمنان رایج که معمولاً خطر کمی دارند.
+کشتن دشمن متوسط (Orc, Skeleton, etc.)		20 XP		دشمنانی که قدرت بیشتری دارند.
+کشتن دشمن قوی (Troll, Dragon, etc.)		100 XP		دشمنان کمیاب و بسیار خطرناک.
+عبور از یک طبقه جدید		50 XP		تشویق برای اکتشاف دالان‌های عمیق‌تر.
+پیدا کردن یک اتاق مخفی		30 XP		تشویق به جستجو در محیط.
+استفاده موفق از اسکرول (مثل شناسایی یا تلپورت)		10 XP		آموزش بازیکن به استفاده از آیتم‌ها.
+خنثی‌سازی یک تله		25 XP		تله‌ها چالش‌های خطرناکی هستند و نیاز به دقت دارند.
+جمع‌آوری یک آیتم کمیاب (Artifact, Special Weapon)		75 XP		بازیکن را تشویق به جستجوی آیتم‌های خاص می‌کند.
+رسیدن به طبقه نهایی		500 XP		جایزه‌ای برای رسیدن به هدف نهایی.
+پیروزی و خروج از سیاه‌چال		1000 XP		بیشترین پاداش برای تمام کردن بازی.
+
